Wiki Entry model with timestamps
