+++ wiki_service/routes/dependencies.py
@@ -1,0 +1,4 @@
+from starlette.requests import Request
+
+def get_db(request: Request):
+    return request.app.state.db
